decouple ParametersMerger
